Move main to cmd
